Added TorchExtension to FSIsolver/extension_operator/extension.py with added initialization with string to model dir.
